+++ sdk/python/feast/infra/offline_stores/file.py
@@ -29,14 +29,14 @@ from feast.infra.offline_stores.offline_utils import (
     DEFAULT_ENTITY_DF_EVENT_TIMESTAMP_COL,
     get_pyarrow_schema_from_batch_source,
 )
-from feast.infra.provider import (
-    _get_requested_feature_views_to_features_dict,
-    _run_dask_field_mapping,
-)
 from feast.registry import BaseRegistry
 from feast.repo_config import FeastConfigBaseModel, RepoConfig
 from feast.saved_dataset import SavedDatasetStorage
 from feast.usage import log_exceptions_and_usage
+from feast.utils import (
+    _get_requested_feature_views_to_features_dict,
+    _run_dask_field_mapping,
+)
 
 
 class FileOfflineStoreConfig(FeastConfigBaseModel):
