+++ sdk/python/tests/utils/online_write_benchmark.py
@@ -14,9 +14,9 @@ from feast.entity import Entity
 from feast.feature_store import FeatureStore
 from feast.feature_view import FeatureView
 from feast.field import Field
-from feast.infra.provider import _convert_arrow_to_proto
 from feast.repo_config import RepoConfig
 from feast.types import Float32, Int32
+from feast.utils import _convert_arrow_to_proto
 
 
 def create_driver_hourly_stats_feature_view(source):
